Change the source for triangulation request
StageListner was a bridge which is redanant
